ok
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@5129 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/libBBS/stuff.c
@@ -1951,9 +1951,18 @@ void long2byte(unsigned int num, byte* arg) {
 
 int my_unlink(char *fname)
 {
+
 #ifndef MYUNLINK_BACKUPDIR
-	return unlink(fname);
+    struct stat st;
+    if (stat(fname, &st))
+        return -1;
+
+    if (!S_ISDIR(st.st_mode))
+        return unlink(fname);
+
+    return rm_dir(fname);
 #else
+
 	char *buf;
 	char *cmd;
 	char *c;
